NewKeyPool support for int64 number of keys.
nIndex is int64, nKeys is int64, it would be useless to have i an int.

+++ src/wallet.cpp
@@ -1412,7 +1412,7 @@ bool CWallet::NewKeyPool()
             return false;
 
         int64 nKeys = max(GetArg("-keypool", 100), (int64)0);
-        for (int i = 0; i < nKeys; i++)
+        for (int64 i = 0; i < nKeys; i++)
         {
             int64 nIndex = i+1;
             walletdb.WritePool(nIndex, CKeyPool(GenerateNewKey()));
